+++ src/client/src/routes/(app)/students/+page.server.ts
@@ -2,6 +2,9 @@ import type { Actions } from './$types';
 import { fail, error, redirect } from '@sveltejs/kit';
 import { generateUsername } from '$utils/user';
 import User from '$db/schemas/User';
+import Workspace from '$db/schemas/Workspace';
+//import { Workspace } from '../../../../../server/workspace/dist/schemas/workspace.schema.js';
+import { orgID } from '../../../lib/store/index.ts';
 
 export async function load({ locals }) {
 	//handle user
@@ -14,8 +17,19 @@ export async function load({ locals }) {
 		const students= await User.find({
 			role: 'student',
 			organisation: locals.user.organisation
-		}).select('_id name surname username email image');
+		}).select('name surname username email image');
+
+		const workspaces= await Workspace.find({
+			organisation: locals.user.organisation
+		}).select('name image');
 	
+		const formattedWorkspaces=workspaces.map((workspace)=>(
+		{
+			id:workspace._id.toString(),
+			name: workspace.name,
+			image: workspace.image
+		}));
+
 
 		const formatStudents= students.map((student)=>(
 			{
@@ -27,7 +41,8 @@ export async function load({ locals }) {
 				image: student.image
 			}));
 		return{
-			students: formatStudents
+			students: formatStudents,
+			workspaces: formattedWorkspaces
 		};
 	}
 	catch (error){
@@ -131,5 +146,34 @@ export const actions: Actions ={
 			console.error('Error removing student:', err);
 			return fail(500, { error: 'Failed to remove student' });
 		}
-	}
+	},
+	enrol: async ({request, locals})=>{
+		if (!locals.user || locals.user.role !== 'admin') throw error(401, 'Unauthorized');
+
+		 // get information from form data
+		 const data = await request.formData();
+		 
+		 const studentId = data.get('id') as string;
+		 const selectedWorkspaces = data.getAll('workspaces') as string[];
+		
+		 try {
+			// Find the student
+			const student = await User.findById(studentId);
+			if (!student) {
+			  return fail(404,{error: 'Student not found'});
+			}
+	  
+			// Update student's workspaces
+			student.workspaces = selectedWorkspaces;
+			await student.save();
+			console.log('Student enrolled successfully');
+	  
+			return {
+			  success: true
+			};
+		  } catch (err) {
+			console.log('Server error:', err);
+			return fail(500, {error: 'Failed to enrol student'});
+		  }
+		}
 }
